asset list renders from db, now to select
