+++ App.js
@@ -7,6 +7,7 @@ import MainNavigation from './navigation/MainNavigation';
 import store from './store';
 
 
+
 export default class App extends Component {
   render() {
